Combine model classes in the kindlegen package into a single file
Trying to keep the package tidy with fewer files.
