+++ src/main/scala/com/gu/kindlegen/Manifests.scala
@@ -2,6 +2,8 @@ package com.gu.kindlegen
 
 import java.time.Instant
 
+import com.gu.io.Link
+
 
 object ArticlesManifest {
   def apply(section: BookSection, buildInstant: Instant = Instant.now): RssManifest = {
@@ -11,6 +13,18 @@ object ArticlesManifest {
       buildInstant = buildInstant,
       publicationDate = section.publicationDate,
       items = section.articles.map(article => RssItem(article.title, article.link))
+    )
+  }
+}
+
+object SectionsManifest {
+  def apply(title: String, link: Link, books: Seq[BookSection], buildInstant: Instant = Instant.now): RssManifest = {
+    RssManifest(
+      title = title,
+      link = link,
+      buildInstant = buildInstant,
+      publicationDate = books.head.publicationDate,
+      items = books.map(_.section).map(section => RssItem(section.title, section.link))
     )
   }
 }
