--- src/main/scala/com/gu/kindlegen/Section.scala
@@ -1,6 +1,0 @@
-package com.gu.kindlegen
-
-import com.gu.io.Link
-
-
-case class Section(id: String, title: String, link: Link)
